requirements.txt
flask
line-bot-sdk==1.18.0
openai
Gunicorn
